:truck: [#516] Move tests from serialiser to endpoint
