+++ backend/src/openarchiefbeheer/destruction/tests/test_serializers.py
@@ -758,129 +758,6 @@ class DestructionListSerializerTests(TestCase):
 
 
 class DestructionListReviewSerializerTests(TestCase):
-    def test_if_user_not_assigned_cannot_create_review(self):
-        reviewer1 = UserFactory.create(
-            username="reviewer1",
-            email="reviewer1@oab.nl",
-            post__can_review_destruction=True,
-        )
-        reviewer2 = UserFactory.create(
-            username="reviewer2",
-            email="reviewer2@oab.nl",
-            post__can_review_destruction=True,
-        )
-        destruction_list = DestructionListFactory.create(assignee=reviewer1)
-
-        data = {
-            "destruction_list": destruction_list.uuid,
-            "decision": ReviewDecisionChoices.accepted,
-        }
-        request = factory.get("/foo")
-        request.user = reviewer2
-
-        serializer = DestructionListReviewSerializer(
-            data=data, context={"request": request}
-        )
-
-        self.assertFalse(serializer.is_valid())
-        self.assertEqual(
-            serializer.errors["author"][0],
-            _(
-                "This user is not currently assigned to the destruction list, "
-                "so they cannot create a review at this stage."
-            ),
-        )
-
-    @override_settings(LANGUAGE_CODE="en")
-    def test_if_list_in_wrong_status_cannot_be_reviewed(self):
-        reviewer = UserFactory.create(
-            username="reviewer",
-            email="reviewer@oab.nl",
-            post__can_review_destruction=True,
-            post__can_review_final_list=True,
-        )
-        destruction_list = DestructionListFactory.create(
-            assignee=reviewer, status=ListStatus.changes_requested
-        )
-
-        data = {
-            "destruction_list": destruction_list.uuid,
-            "decision": ReviewDecisionChoices.accepted,
-        }
-        request = factory.get("/foo")
-        request.user = reviewer
-
-        serializer = DestructionListReviewSerializer(
-            data=data, context={"request": request}
-        )
-
-        self.assertFalse(serializer.is_valid())
-        self.assertEqual(
-            serializer.errors["author"][0],
-            _(
-                "The status of this destruction list prevents you from creating a review at this stage."
-            ),
-        )
-
-    @override_settings(LANGUAGE_CODE="en")
-    def test_if_user_not_a_reviewer_cannot_create_review(self):
-        reviewer = UserFactory.create(
-            username="reviewer",
-            email="reviewer@oab.nl",
-            post__can_review_destruction=False,
-        )
-        destruction_list = DestructionListFactory.create(
-            assignee=reviewer, status=ListStatus.ready_to_review
-        )
-
-        data = {
-            "destruction_list": destruction_list.uuid,
-            "decision": ReviewDecisionChoices.accepted,
-        }
-        request = factory.get("/foo")
-        request.user = reviewer
-
-        serializer = DestructionListReviewSerializer(
-            data=data, context={"request": request}
-        )
-
-        self.assertFalse(serializer.is_valid())
-        self.assertEqual(
-            serializer.errors["author"][0],
-            _(
-                "The status of this destruction list prevents you from creating a review at this stage."
-            ),
-        )
-
-    @override_settings(LANGUAGE_CODE="en")
-    def test_if_user_not_an_archivist_cannot_create_review(self):
-        reviewer = UserFactory.create(
-            username="reviewer",
-            email="reviewer@oab.nl",
-            post__can_review_final_list=False,
-        )
-        destruction_list = DestructionListFactory.create(
-            assignee=reviewer, status=ListStatus.ready_for_archivist
-        )
-
-        data = {
-            "destruction_list": destruction_list.uuid,
-            "decision": ReviewDecisionChoices.accepted,
-        }
-        request = factory.get("/foo")
-        request.user = reviewer
-
-        serializer = DestructionListReviewSerializer(
-            data=data, context={"request": request}
-        )
-
-        self.assertFalse(serializer.is_valid())
-        self.assertEqual(
-            serializer.errors["author"][0],
-            _(
-                "The status of this destruction list prevents you from creating a review at this stage."
-            ),
-        )
 
     def test_create_review_accepted(self):
         reviewer = UserFactory.create(
